+++ src/Utils/UserIdUtil.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+ * Copyright 2024-2025 Wingify Software Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace vwo\Utils;
+
+use vwo\Services\SettingsService;
+use vwo\Services\LoggerService;
+
+class UserIdUtil
+{
+  /**
+   * Resolves the canonical userId considering aliasing feature and gateway availability
+   * @param string $userId
+   * @param bool $isAliasingEnabled
+   * @return string
+   */
+  public static function getUserId($userId, $isAliasingEnabled)
+  {
+    if ($isAliasingEnabled) {
+      if (SettingsService::instance()->isGatewayServiceProvided) {
+        $aliasId = AliasingUtil::getAlias($userId);
+
+        LoggerService::info('ALIAS_ENABLED', ['userId' => $aliasId]);
+        return $aliasId;
+      } else {
+        LoggerService::error('GATEWAY_URL_ERROR');
+        return $userId;
+      }
+    } else {
+      return $userId;
+    }
+  }
+}
+
+?>
+
+